requirements.txt
pygame
pyinstaller
pillow
numpy